Test harness for rust verification results
This runs the compiler in the test process, on files in memory, and extracts span information for the verification errors, so we don’t have to rely on the output format.

The idea here is that you get line numbers and error descriptions, but you also get the entire line of the span, which lets you leave a marker in a comment, and check that the reported error is on the line you expected (without having to compute line numbers).