Add an "advanced" macro transformer
This macro transformer restricts the usage of its transformations to
only scenarios where it is explicitly escaped (currently, this means a
token tree must be prefixed with `@@` to be transformed from Verus
syntax). If something is not prefixed with the `@@` then it is left
untransformed.

Note that this current approach only applies potential transformation at
the root level of the expr, and thus cannot be used like (say) lisp's
unquote operator (the `,` inside a quoted expression). However, this
_may_ be something we may wish to add in the future at some point?