A few additions to pervasive for vec, seq, and set (#472)
Adds insert and remove for sequences, and uses them to add equivalent operations on vec.
Adds a utility definition for is_full() for set, since using a direct comparison with Set::full() requires a type annotation. It also it provides a parallel for is_empty().
Adds a utility lemma for mapping sequences of unique items to sets, building on @achreto's lemmas.