zsh + nix support on `main_new` (#483)
* Support shell activation in zsh too

Note that this is needed for this script (but not necessary in most
other scripts), since most other scripts can actually be invoked via
`bash` (and indeed, their `#!` does just that), but this particular
script is intended to be `source`d in, which means that the specific
shell that is doing the `source`ing is what is in charge of running the
whole script, and the interpreter-line `#!` is ignored. Thus, we need to
explicitly account for zsh too.

* Remove the `#!` for script intended to be `source`d

* Update shell.nix with latest build requirements

Turns out, in the past, we needed a lot more specific build requirements
due to us using a fork of Rust. Now that we are on mainline Rust, all we
really need rustup and a tiny number of utilities