rust_verify: record and report time on a per-module basis
with the parallel verification, the time reporting needed adaption.
we now report the real time for rust, verification, and the unaccounted
parts. We further report a breakdown of the total parallel verification
time with a breakdown per-module. Finally, we give an estimate of the
total CPU time, i.e., when running verification using a single thread.

Signed-off-by: Reto Achermann <achreto@gmail.com>