fix type-inference issues in ghost code
by using a quirk of type inference with (inline) functions and
`unreachable_unchecked` returning never.

In a nutshell, we're going to exploit this behavior:

```rust
pub struct Ghost<A> {
    p: std::marker::PhantomData<A>,
}

impl<A> Ghost<A> {
    fn from_fn(_: fn()->A) -> Ghost<A> {
        Ghost { p: std::marker::PhantomData }
    }
}

pub fn foo() {
    let l = Ghost::from_fn(|| unreachable!());
}
```

This type-checks successfully, despite `l` appearing underconstrained.

This is exactly the situation we encounter in a program like:

```rust
fn main() {
    let b = Ghost(4nat);
}
```

where `b` remained unconstrained after erasure.

This example program used to fail during compilation with:

```rust
error[E0282]: type annotations needed for `builtin::Ghost<A>`
  |
6 |     let b = Ghost(4nat);
  |         ^
  |
help: consider giving `b` an explicit type, where the type for type parameter `A` is specified
  |
6 |     let b: builtin::Ghost<A> = Ghost(4nat);
  |          +++++++++++++++++++

error: aborting due to previous error
```

and it now compiles successfully with this patch.

The technique was discovered in collaboration with @xldenis.