Fix the rustup toolchain channel regex
The regex used to extract the channel from the output of `rustup show
active-toolchain` misidentifies `aarch64` as part of the channel on Linux:

```
$ vargo build
error: rustup is using a toolchain with channel 1.76.0-aarch64, we expect 1.76.0
do you have a rustup override set?
$ rustup show active-toolchain
1.76.0-aarch64-unknown-linux-gnu (overridden by '/verus/rust-toolchain.toml')
```

This is the result of a regex that consumes `-` as part of the channel name:
`^(([A-Za-z0-9.-]+)-[A-Za-z0-9_]+-[A-Za-z0-9]+-[A-Za-z0-9-]+)`

The first capture group can't contain `_`, so it works on `x86_64`.  According
to the [toolchain spec
grammar](https://rust-lang.github.io/rustup/concepts/toolchains.html#toolchain-specification),
we'll need to use a less liberal regex to avoid capturing too much as the
channel/date.

```
<channel>[-<date>][-<host>]

<channel>       = stable|beta|nightly|<major.minor>|<major.minor.patch>
<date>          = YYYY-MM-DD
<host>          = <target-triple>
```

The regex also causes problems on platforms that use a target "triple" that
isn't a triple, like `1.76.0-aarch64-apple-darwin`, so we make no assumptions
about the number of `-`-separated words in the target triple.