silence superfluous warnings
When using Cargo to build a Rust/Verus project that imports vstd using a
local path (instead of a git repo), cargo doesn't silence warnings from
building various packages from the verus repo, and there's a lot of them.
This commit silences all of the warnings (all of which seem benign).