Add new badge
It came to my attention that some folks might find it helpful to have a badge that points at _Verus_ docs (I know we mention the guide later in the README under docs, but this is more of a direct jump to it). One _can_ reach it from the tutorial shield, but I think it is still helpful to have an extra badge for the reference, that tells people that docs about Verus (in contrast with vstd docs) are here. Some folks end up accidentally jumping to vstd docs and end up a little confused, and then need to navigate to the tutorial from there.