Add one prelude axiom to be identified in `axiom-usage-info` (#1397)
The proof of concrete sequence sorted by mergesort requires a prelude axiom, which is not reflected by verus -V broadcast-usage-info before, as the verus will only take the named axioms in unsat-core. We named one axiom in air prelude (which seems to be the only one not about unboxing/boxing conversions or has_type).

We also changed broadcast-usage-info to axiom-usage-info to match the vargo build option (vargo build --features axiom-usage-info)

---

Co-authored-by: Andrea Lattuada <andrea@mpi-sws.org>