Leverage Cargo (continued) (#1475)
* Fix examples

* Add rustfmt::skip attribute to vstd modules

To prepare for including vstd in the workspace. Once it is included in
the workspace, it will be visited by cargo fmt.

We guard the rustfmt::skip attributes with:

'#![cfg_attr(rustfmt, rustfmt::skip)]'

because custom inner attributes are unstable.

* Add cargo-verus tool and split out verus-driver

* REMOVE BEFORE MERGING: Add cargo-verus demo

Just to aid discussion and review.

* Ensure cargo-verus and verus-driver versions align

* cargo-verus: Relax verus-driver version parsing

* verus-driver: Improve handling of -h and -V

* verus-driver: Improve passthru detection

* cargo-verus: Distinguish cargo-verus output

This trick makes uses of a mechanism designed for building the Rust
sysroot to ensure that the output of verus-driver via cargo-verus is
distinguished from the output of rustc.

The value of the __CARGO_DEFAULT_LIB_METADATA is incorporated into the
hashes that are included in compilation output filenames.

* cargo-verus: Clean up temporary help message

For until a proper one is added.

* verus-driver: Clean up temporary error message

For until a proper one is added.

* cargo-verus: Automatically detect VIR imports

Using the package.metadata.verus tables of dependencies.

* Refactor and improve behavior

* cargo-verus: Index cargo metadata output

* Refactor for clarity

* cargo-verus: Improve error handling

Using the 'anyhow' crate.

* verus-driver: Improve error handling

Using 'rustc_session::EarlyDiagCtxt'.

* Undo changes to Cargo.lock and some formatting-related changes

* Move some files

* Move catch_with_exit_code and re-indent resulting code

* Integrate verus-driver functionality directly into rust_verify

* Point cargo-verus's RUSTC_WRAPPER to verus executable

* Remove obsolete --cfg erasure_macro_todo

* Update import_crates for cargo verus

* Add cargo-verus to vargo

* Remove cargo-verus-illustration directory

* Update example/cargo-verus/run.sh

* Rename example to examples

* Change command-line syntax to check/verify/build

* Remove some unused code

* more liberal parsing of the active toolchain in vargo and the verus build script

* move the `via_cargo` args out of the main config path

---------

Co-authored-by: Nick Spinale <nick@nickspinale.com>
Co-authored-by: Andrea Lattuada <andrea@mpi-sws.org>