fix overly-restrictive case in mask subset checking (#1586)
The invariant mask-checking code was buggy (in a way that was sound but overly restrictive). In particular, the equivalent of checking s.subset_eq(Set::full().remove(i)) always failed (false), but it should be allowed (true) if !s.contains(i). Checking the latter was the intent of the code, but the code was buggy. This fixes the bug and introduces some test cases to make sure this case is handled correctly.