new-mut-ref: building blocks for general mut ref support
Based on the initial work in #1738, we make `&mut T` a "first class" type and encode
it into AIR as a pair. However, we further generalize the support for mut_refs beyond
arguments to function calls by creating new VIR nodes for mutable borrows, separate from
the calls, along with other relevant functionality.

Specifically we add the following to VIR:

 - Explicit `BorrowMut` nodes
 - `mut_ref_current`, `mut_ref_future`, and `has_resolved` spec functions
 - A "resolve" operation

The functionality in this PR is incomplete, and only begins the new-mut-ref effort.
All new functionality is beyond a development flag `-V new-mut-ref`.

Co-authored-by: Andrea Lattuada <andrea@mpi-sws.org>
Co-authored-by: Abigail Pribisova <apribiso@mpi-sws.org>