bugfix: recommends-failure when returns clause is involved; fixes #1834
The `ens_has_return` flag was intended for exec/proof code, but it
turns out that setting it to `false` on a spec function will cause
recommends-checking to break. The problem was that the synthesized 'spec'
function to go with a 'returns' clause was setting the `ens_has_return`
to `false`. The fix here is simply to set it to `true`.