Update the existing GitHub release workflow to also publish any updated crates to crates.io (#1881)
* Add ability to invoke `cargo publish`
* Update default features for verus_syn, so that a bare `cargo publish`
succeeds.  Update verus_prettyplease to specify the full set of features
it needs from verus_syn, so that a bare `cargo publish` succeeds with it
as well.
* Update the release workflow to run the version bumper and publish any
updated Verus crates
* Offer a --publish instead of --dry-run option, as per @jaybosamiya-ms's
suggestion.
* Update the version bumper to bump versions for all directly modified
crates *and* the crates that transitively depend on them.
* Swap to a subcommand style, so that we can externally perform
a git commit between the update and publish steps.
* Update the release workflow to update crate versions, git commit the
changes, and publish to crates.io.  Everything is still in practice
mode, aside from the git changes.