map: add lemmas around `Map::values` lengths
This commit adds two lemmas `Map::lemma_values_len` and
`Map::lemma_injective_values_len` to reason about the length of
`m.values().len()` (in particular that `m.values().len() <=
m.dom().len()`) and the fact that this reaches equality for the case if
the map is injective.

This required a weakening of the precondition on `Set::lemma_map_size`.
Previously, this lemma required the mapping function to be injective everywhere, even
outside the mapping domain. Now, it does not, matching the Dafny
function this was based on. Note that:

1. Sometimes it is impossible for a function to be injective in the
   mapping domain (if the mapped to type has a lower cardinality);
2. It is not necessary -- having the weaker precondition changes nothing
   about the original proof.