ci: refactor tests for clarity
We effectively have three tiers of support:
1. MacOs-arm (check that builds even with esoteric options and runs all tests)
2. Linux, Windows, MacOs-x86-64 (check that builds and passes basic
   tests)
3. Esoteric builds (no-std/no-alloc/singular/record-history): check that builds and,
   in the case of singular and record history, that it passes a couple
   of tests

This commit makes the difference a bit more clear.

Tiers 1/2 now do the `test` job, where they are fully built. These are
also the releases we offer.

Tier 3 is built and runs the singular/record history tests.

doc script flag was changed for clarity (and we skip just checking the
docs and then rebuilding them in another job).