ci: refactor for maintainability and optimize for latency (#2024)
* ci: update rust toolchain setup

* ci: breakup and simplify workflow + run tests in release mode

* ci: move from cargo test to cargo nextest

* vargo: respect user provided RUSTFLAGS

* ci: enable debug assertions in release tests

* ci: revert building tests with release and separate building artifacts

* ci: move to latest macos

* ci: refactor tests for clarity

We effectively have three tiers of support:
1. MacOs-arm (check that builds even with esoteric options and runs all tests)
2. Linux, Windows, MacOs-x86-64 (check that builds and passes basic
   tests)
3. Esoteric builds (no-std/no-alloc/singular/record-history): check that builds and,
   in the case of singular and record history, that it passes a couple
   of tests

This commit makes the difference a bit more clear.

Tiers 1/2 now do the `test` job, where they are fully built. These are
also the releases we offer.

Tier 3 is built and runs the singular/record history tests.

doc script flag was changed for clarity (and we skip just checking the
docs and then rebuilding them in another job).

* ci: update ci README